Working on enemy types
Tried implementing an eel enemy that attacks from the bottom of the screen and then goes away, but the logic to notice when it gets to the players y position, and then retreat if it missed the player isn't working. Not sure what the disconnect is between the two top position values at the moment.
I'll probably abandon this as it won't likely be possible before the submission deadline.